WIP: duplicate functionality for dashboard
