+++ main/index.js
@@ -57,7 +57,12 @@ if (!is.development) {
 app.whenReady().then(() => {
   loadMenu(true)
   // TODO: not necesarily latest...
-  const latestFile = knownFilesStore.store[0]
+  const files = Object.values(knownFilesStore.store).sort((thisFile, thatFile) => {
+    if (thisFile.lastOpened > thatFile.lastOpened) return 1
+    if (thisFile.lastOpened < thatFile.lastOpened) return -1
+    return 0
+  })
+  const latestFile = files[0]
   openProjectWindow(latestFile.path)
   windowsOpenFileEventHandler(process.argv)
