+++ main/modules/broadcast.js
@@ -1,0 +1,7 @@
+const { BrowserWindow } = require('electron')
+
+export const broadcastToAllWindows = (event, payload) => {
+  BrowserWindow.getAllWindows().forEach((bw) => {
+    bw.webContents.send('set-dark-mode', payload)
+  })
+}
