+++ main/modules/dashboard.js
@@ -1,0 +1,5 @@
+const { broadcastToAllWindows } = require('./broadcast')
+
+export const reloadRecents = () => {
+  broadcastToAllWindows('reload-recents')
+}
