+++ main/modules/menus/dashboard.js
@@ -1,0 +1,5 @@
+import { ipcRenderer } from 'electron'
+
+export const openDashboard = () => {
+  ipcRenderer.send('open-dashboard')
+}
