+++ main/modules/menus/file.js
@@ -4,8 +4,8 @@ const { app, shell } = require('electron')
 const { is } = require('electron-util')
 const { getWindowById, numberOfWindows } = require('../windows')
 const { NODE_ENV } = require('../constants')
-const { openDashboard } = require('../windows/dashboard')
 const { newFileOptions } = require('../new_file_options')
+const { openDashboard } = require('./dashboard')
 
 const TEMP_FILES_PATH = path.join(app.getPath('userData'), 'tmp')
 let showInMessage = i18n('Show in File Explorer')
