+++ main/modules/windows/files.js
@@ -1,0 +1,4 @@
+export const updateOpenFiles = () => {
+  // TODO!  Update the list of open files so that other dashboards
+  // know which files are open.
+}
