+++ main/modules/windows/projects.js
@@ -3,7 +3,7 @@ const { app, ipcMain } = require('electron')
 const log = require('electron-log')
 const { makeBrowserWindow } = require('../utils')
 const { filePrefix } = require('../helpers')
-const { updateOpenFiles } = require('./dashboard')
+const { updateOpenFiles } = require('./files')
 const { rollbar } = require('../rollbar')
 // const { NODE_ENV } = require('../constants')
 const { getWindowById, addNewWindow, dereferenceWindow, focusIfOpen } = require('.')
